pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "le_professeur_bizarre"
version = "1.0.0"
description = "A quirky Franco-American cultural teacher for Reachy Mini - translates English to French with bizarre cultural facts!"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Francisco", email = "francisco@example.com"}
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "reachy-mini>=1.2.0",
    "httpx>=0.24.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]
mujoco = [
    "reachy-mini[mujoco]>=1.2.0",
]

[project.urls]
Homepage = "https://huggingface.co/spaces/Franciscomoney/le_professeur_bizarre"
Repository = "https://github.com/Franciscomoney/le_professeur_bizarre"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
"le_professeur_bizarre" = ["static/*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
